refactor(cms-sync): extract asset processor, deduplicate check-cms-assets
