+++ scripts/check-cms-assets.js
@@ -11,6 +11,7 @@ const fs = require('fs')
 const path = require('path')
 const { execSync } = require('child_process')
 const matter = require('gray-matter')
+const { extractAssetPaths } = require('./cms-sync/asset-processor')
 
 const SYNC_FOLDERS = [
   'faqs',
@@ -23,72 +24,6 @@ const SYNC_FOLDERS = [
 ]
 
 const MIGRATED_PATTERN = new RegExp(`^data/(${SYNC_FOLDERS.join('|')})/.*\\.mdx?$`)
-
-function stripFencedCodeBlocks(content) {
-  let out = content.replace(/^[ \t]*```[^\n]*\n[\s\S]*?^[ \t]*```/gm, '\n')
-  out = out.replace(/^[ \t]*~~~[^\n]*\n[\s\S]*?^[ \t]*~~~/gm, '\n')
-  return out
-}
-
-function extractAssetPaths(content, frontmatter) {
-  const paths = new Set()
-  const bodyForScan = stripFencedCodeBlocks(content)
-
-  // Markdown images: ![alt](/path/to/image.webp)
-  const mdImageRegex = /!\[.*?\]\((.*?)\)/g
-  let match
-  while ((match = mdImageRegex.exec(bodyForScan)) !== null) {
-    if (match[1] && !match[1].startsWith('http') && !match[1].startsWith('https')) {
-      paths.add(match[1])
-    }
-  }
-
-  // JSX/HTML component src attributes
-  const componentTags = ['img', 'video', 'source', 'Image', 'Figure', 'Table', 'NextImage']
-  componentTags.forEach((tagName) => {
-    const tagRegex = new RegExp(
-      `<${tagName}[^>]*?\\s+src\\s*=\\s*["']([^"']+)["'][^>]*?(?:/>|>[\\s\\S]*?</${tagName}>)`,
-      'gi'
-    )
-    let m
-    while ((m = tagRegex.exec(bodyForScan)) !== null) {
-      if (m[1] && !m[1].startsWith('http') && !m[1].startsWith('https')) {
-        paths.add(m[1])
-      }
-    }
-
-    const tagRegexNoQuotes = new RegExp(
-      `<${tagName}[^>]*?\\s+src\\s*=\\s*([^\\s>"']+)[^>]*?(?:/>|>[\\s\\S]*?</${tagName}>)`,
-      'gi'
-    )
-    while ((m = tagRegexNoQuotes.exec(bodyForScan)) !== null) {
-      if (m[1] && !m[1].startsWith('http') && !m[1].startsWith('https')) {
-        paths.add(m[1])
-      }
-    }
-  })
-
-  // Frontmatter fields with local asset paths
-  function checkValue(value) {
-    if (typeof value === 'string') {
-      if (
-        value.startsWith('/') &&
-        !value.startsWith('http') &&
-        !value.startsWith('https') &&
-        /\.[a-zA-Z0-9]+$/.test(value)
-      ) {
-        paths.add(value)
-      }
-    } else if (Array.isArray(value)) {
-      value.forEach(checkValue)
-    } else if (typeof value === 'object' && value !== null) {
-      Object.values(value).forEach(checkValue)
-    }
-  }
-
-  checkValue(frontmatter)
-  return Array.from(paths)
-}
 
 function getStagedMigratedFiles() {
   const output = execSync('git diff --cached --name-only --diff-filter=ACM', { encoding: 'utf8' })
